Refactored names and migration (1-5).
