+++ pom.xml
@@ -112,7 +112,7 @@
                 <version>8.0.5</version>
                 <configuration>
                     <url>${env.ENDABANK_DB_URL}</url>
-                    <user>${env.ENDABANK_DB_USER}</user>
+                    <user>${env.ENDABANK_DB_USERNAME}</user>
                     <password>${env.ENDABANK_DB_PASSWORD}</password>
                 </configuration>
             </plugin>
