+++ src/main/java/com/endava/endabank/controller/AuthenticationController.java
@@ -1,4 +1,4 @@
-package com.endava.endabank.controllers;
+package com.endava.endabank.controller;
 
 import com.endava.endabank.constants.Routes;
 import com.endava.endabank.dto.user.AuthenticationDto;
@@ -8,7 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,8 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping(Routes.API_ROUTE)
