+++ src/main/java/com/endava/endabank/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.endava.endabank.controllers;
+package com.endava.endabank.controller;
 
 import com.endava.endabank.constants.Permissions;
 import com.endava.endabank.constants.Routes;
@@ -6,7 +6,7 @@ import com.endava.endabank.dto.user.UpdatePasswordDto;
 import com.endava.endabank.dto.user.UserPrincipalSecurity;
 import com.endava.endabank.dto.user.UserRegisterDto;
 import com.endava.endabank.exceptions.customExceptions.ActionNotAllowedException;
-import com.endava.endabank.services.UserService;
+import com.endava.endabank.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
