+++ src/main/java/com/endava/endabank/dao/IdentifierTypeDao.java
@@ -1,6 +1,6 @@
 package com.endava.endabank.dao;
 
-import com.endava.endabank.models.IdentifierType;
+import com.endava.endabank.model.IdentifierType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IdentifierTypeDao extends JpaRepository<IdentifierType, Integer> {
