+++ src/main/java/com/endava/endabank/dao/RoleDao.java
@@ -1,6 +1,6 @@
 package com.endava.endabank.dao;
 
-import com.endava.endabank.models.Role;
+import com.endava.endabank.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleDao extends JpaRepository<Role,Integer> {
