+++ src/main/java/com/endava/endabank/model/Permission.java
@@ -1,37 +1,34 @@
-package com.endava.endabank.models;
+package com.endava.endabank.model;
 
 import lombok.*;
 import org.hibernate.Hibernate;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 @Entity
-@Table(name = "state_types")
+@Table(name = "permissions")
 @Getter
 @Setter
 @ToString
 @RequiredArgsConstructor
-public class StateType {
+public class Permission {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(nullable = false)
     private Integer id;
     @Column(length = 40, nullable = false)
     private String name;
-
-    @OneToMany(mappedBy = "stateType", cascade = CascadeType.ALL, orphanRemoval = true)
+    @ManyToMany(cascade = CascadeType.ALL, mappedBy = "permissions")
     @ToString.Exclude
-    private List<TransactionState> transactionStates = new ArrayList<>();
+    private List<Role> roles;
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        StateType stateType = (StateType) o;
-        return id != null && Objects.equals(id, stateType.id);
+        Permission that = (Permission) o;
+        return id != null && Objects.equals(id, that.id);
     }
 
     @Override
