+++ src/main/java/com/endava/endabank/model/StateType.java
@@ -1,35 +1,37 @@
-package com.endava.endabank.models;
+package com.endava.endabank.model;
 
 import lombok.*;
 import org.hibernate.Hibernate;
 
 import javax.persistence.*;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 @Entity
-@Table(name = "transaction_types")
+@Table(name = "state_types")
 @Getter
 @Setter
 @ToString
 @RequiredArgsConstructor
-public class TransactionType {
+public class StateType {
     @Id
-    @Column(nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(nullable = false)
     private Integer id;
-    @Column(length = 20, nullable = false)
-    private  String name;
-    @OneToMany(mappedBy = "transactionType")
+    @Column(length = 40, nullable = false)
+    private String name;
+
+    @OneToMany(mappedBy = "stateType", cascade = CascadeType.ALL, orphanRemoval = true)
     @ToString.Exclude
-    private List<Transaction> transactions;
+    private List<TransactionState> transactionStates = new ArrayList<>();
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        TransactionType that = (TransactionType) o;
-        return id != null && Objects.equals(id, that.id);
+        StateType stateType = (StateType) o;
+        return id != null && Objects.equals(id, stateType.id);
     }
 
     @Override
