+++ src/main/resources/db/migration/V11__edit_column_amount.sql
@@ -1,0 +1,5 @@
+ALTER TABLE transactions ADD amount FLOAT;
+
+ALTER TABLE transactions ALTER COLUMN  amount SET NOT NULL;
+
+ALTER TABLE transactions DROP COLUMN ammount;
